requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
requests
python-dotenv
httpx
supabase-py
passlib
auth0-python
